# src/main/java/com/zufar/icedlatte/security/endpoint/UserSecurityEndpoint.java
package com.zufar.icedlatte.security.endpoint;

import com.zufar.icedlatte.email.api.*;
import com.zufar.icedlatte.openapi.dto.*;
import com.zufar.icedlatte.openapi.security.api.SecurityApi;
import com.zufar.icedlatte.security.api.*;
import com.zufar.icedlatte.security.jwt.*;
import com.zufar.icedlatte.user.api.*;
import jakarta.servlet.http.HttpServletRequest;
import jakarta.validation.Valid;
import lombok.RequiredArgsConstructor;
import lombok.extern.slf4j.Slf4j;
import org.apache.commons.text.StringEscapeUtils;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@Slf4j
@Validated
@RestController
@RequestMapping(value = UserSecurityEndpoint.USER_SECURITY_API_URL)
@RequiredArgsConstructor
public class UserSecurityEndpoint implements SecurityApi {

    public static final String USER_SECURITY_API_URL = "/api/v1/auth/";

    private final UserAuthenticationService userAuthenticationService;
    private final JwtTokenFromAuthHeaderExtractor jwtTokenFromAuthHeaderExtractor;
    private final JwtBlacklistValidator jwtBlacklistValidator;
    private final EmailTokenSender emailTokenSender;
    private final EmailTokenConformer emailTokenConformer;
    private final JwtAuthenticationProvider jwtAuthenticationProvider;
    private final UserDetailsService userDetailsService;
    private final SingleUserProvider singleUserProvider;
    private final ChangeUserPasswordOperationPerformer changeUserPasswordOperationPerformer;

    private final HttpServletRequest httpRequest;

    @PostMapping("/register")
    public ResponseEntity<String> register(@RequestBody @Valid final UserRegistrationRequest request) {
        var email = request.getEmail();
        log.info("Registering user: {}", email);
        emailTokenSender.sendEmailVerificationCode(request);
        log.info("Verification email sent to: {}", email);
        return ResponseEntity.ok(
            String.format("Email verification token sent to the user with email = %s", 
                StringEscapeUtils.escapeJava(email)));
    }

    @Override
    @PostMapping(value = "/confirm")
    public ResponseEntity<UserRegistrationResponse> confirmEmail(@Validated @Valid @RequestBody final ConfirmEmailRequest confirmEmailRequest) {
        log.info("Confirming email verification");
        var response = emailTokenConformer.confirmEmailByCode(confirmEmailRequest);
        log.info("Email verification completed");
        return ResponseEntity.status(HttpStatus.CREATED).body(response);
    }

    @Override
    @PostMapping("/authenticate")
    public ResponseEntity<UserAuthenticationResponse> authenticate(@Valid @RequestBody final UserAuthenticationRequest request) {
        // Validate input to prevent code injection
        if (request.getEmail() == null || request.getEmail().trim().isEmpty() ||
            request.getPassword() == null || request.getPassword().trim().isEmpty()) {
            log.warn("Invalid authentication request: missing credentials");
            return ResponseEntity.badRequest().build();
        }
        
        log.info("Authenticating user: {}", request.getEmail());
        var response = userAuthenticationService.authenticate(request);
        log.info("Authentication completed for: {}", request.getEmail());
        return ResponseEntity.ok(response);
    }

    @Override
    @PostMapping("/refresh")
    public ResponseEntity<UserAuthenticationResponse> refreshToken() {
        log.info("Refreshing token");
        var authToken = jwtAuthenticationProvider.get(httpRequest);
        var userDetails = userDetailsService.loadUserByUsername(authToken.getName());
        var response = userAuthenticationService.authenticate(userDetails, authToken.getName());
        log.info("Token refreshed for: {}", authToken.getName());
        return ResponseEntity.ok(response);
    }

    // import org.springframework.security.web.csrf.CsrfFilter;
    // import org.springframework.security.config.annotation.web.builders.HttpSecurity;

    @Override
    @PostMapping("/logout")
    public ResponseEntity<Void> logout() {
        log.info("Processing logout");
        var token = jwtTokenFromAuthHeaderExtractor.extract(httpRequest);
        if (token != null) {
            jwtBlacklistValidator.addToBlacklist(token);
            log.info("Token blacklisted");
        }
        return ResponseEntity.ok().build();
    }

    @Override
    @PostMapping("/password/forgot")
    public ResponseEntity<Void> forgotPassword(@Valid @RequestBody final ForgotPasswordRequest request) {
        // Validate input to prevent code injection
        if (request.getEmail() == null || request.getEmail().trim().isEmpty()) {
            log.warn("Invalid forgot password request: null or empty email");
            return ResponseEntity.badRequest().build();
        }
        
        log.info("Processing forgot password for: {}", request.getEmail());
        var user = singleUserProvider.getUserByEmail(request.getEmail());
        var verificationRequest = new UserRegistrationRequest(user.getFirstName(), user.getLastName(), user.getEmail(), "");
        emailTokenSender.sendEmailVerificationCode(verificationRequest);
        log.info("Password reset email sent to: {}", request.getEmail());
        return ResponseEntity.ok().build();
    }

    @Override
    @PostMapping("/password/change")
    public ResponseEntity<Void> changePassword(@Valid @RequestBody final ChangePasswordRequest request) {
        // Validate input to prevent code injection
        if (request.getEmail() == null || request.getEmail().trim().isEmpty() ||
            request.getCode() == null || request.getCode().trim().isEmpty() ||
            request.getPassword() == null || request.getPassword().trim().isEmpty()) {
            log.warn("Invalid change password request: missing required fields");
            return ResponseEntity.badRequest().build();
        }
        
        log.info("Changing password for: {}", request.getEmail());
        var user = singleUserProvider.getUserByEmail(request.getEmail());
        emailTokenConformer.confirmResetPasswordEmailByCode(new ConfirmEmailRequest(request.getCode()));
        changeUserPasswordOperationPerformer.changeUserPassword(user.getId(), request.getPassword());
        log.info("Password changed for: {}", request.getEmail());
        return ResponseEntity.ok().build();
    }
}